+++ crates/rust-client/src/builder.rs
@@ -11,7 +11,7 @@ use rand::Rng;
 use crate::keystore::FilesystemKeyStore;
 use crate::note_transport::NoteTransportClient;
 use crate::rpc::NodeRpcClient;
-use crate::store::Store;
+use crate::store::{Store, StoreError};
 use crate::{Client, ClientError, DebugMode};
 
 // CONSTANTS
@@ -35,6 +35,23 @@ enum AuthenticatorConfig<AUTH> {
     Instance(Arc<AUTH>),
 }
 
+// STORE BUILDER
+// ================================================================================================
+
+/// Allows the [`ClientBuilder`] to accept either an already built store instance or a factory for
+/// deferring the store instantiation.
+pub enum StoreBuilder {
+    Store(Arc<dyn Store>),
+    Factory(Box<dyn StoreFactory>),
+}
+
+/// Trait for building a store instance.
+#[async_trait::async_trait]
+pub trait StoreFactory {
+    /// Returns a new store instance.
+    async fn build(&self) -> Result<Arc<dyn Store>, StoreError>;
+}
+
 // CLIENT BUILDER
 // ================================================================================================
 
@@ -47,7 +64,7 @@ pub struct ClientBuilder<AUTH> {
     /// An optional custom RPC client. If provided, this takes precedence over `rpc_endpoint`.
     rpc_api: Option<Arc<dyn NodeRpcClient>>,
     /// An optional store provided by the user.
-    store: Option<Arc<dyn Store>>,
+    pub store: Option<StoreBuilder>,
     /// An optional RNG provided by the user.
     rng: Option<Box<dyn FeltRng>>,
     /// The keystore configuration provided by the user.
@@ -70,7 +87,6 @@ impl<AUTH> Default for ClientBuilder<AUTH> {
             rpc_api: None,
             store: None,
             rng: None,
-
             keystore: None,
             in_debug_mode: DebugMode::Disabled,
             tx_graceful_blocks: Some(TX_GRACEFUL_BLOCKS),
@@ -116,7 +132,7 @@ where
     /// Optionally provide a store directly.
     #[must_use]
     pub fn store(mut self, store: Arc<dyn Store>) -> Self {
-        self.store = Some(store);
+        self.store = Some(StoreBuilder::Store(store));
         self
     }
 
@@ -188,11 +204,14 @@ where
         };
 
         // Ensure a store was provided.
-        let store: Arc<dyn Store> = if let Some(store) = self.store {
-            store
+        let store = if let Some(store_builder) = self.store {
+            match store_builder {
+                StoreBuilder::Store(store) => store,
+                StoreBuilder::Factory(factory) => factory.build().await?,
+            }
         } else {
             return Err(ClientError::ClientInitializationError(
-                "Store must be specified. Call `.store(...)` or `.sqlite_store(...)` with a store path if `sqlite` is enabled."
+                "Store must be specified. Call `.store(...)` or `.sqlite_store(...)` with a store path."
                     .into(),
             ));
         };